Let Strimzi reference its images the old way
Strimzi only pulls kafka/zookeeper once for a limited number of
nodes (3, since we deploy 3 nodes of each) and only on initial cluster
setup or potentially when rebooting the Kafka or Zookeeper pods. But,
we don't ever reboot these under normal operations so let's just let
those come from DockerHub for now.

This should get Travis green again and the Strimzi cluster-controller
looks like it's overriding the image to the DockerHub version anyway.