Temporarily pin js dep for CouchDB until we work out a proper fix
CentOS 7 bumped its `js` version from 1.8.5-19 to 1.8.5-20 and
apparently broke CouchDB's JavaScript view generation in the
process. https://github.com/apache/couchdb/issues/1293 tracks that
issue.

So, for now, we pin to 1.8.5-19. It's not available to install
directly from the repos so we point to a download location on
vault.centos.org.